backend/services/resume_parser/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
boto3
PyMuPDF
python-docx